Resolve Issue #12, Fix hint button
Resolved Issue #12 by fixing the hint button, which was giving out letters
already guessed as hints.

The function to check whether a letter had been used only accepted uppercase
letters while the function to choose a hint letter was passing in lowercase
letters. Changed the hint function to pass in an uppercase letter. As a result,
the hint function now works as intended.

Changing the passed in letter within the letter checking function to uppercase
may provide a better solution while also avoiding this issue in the future

Resolves: #12